Move LollipopCheckStatus under types folder
