+++ ts/features/lollipop/types/LollipopCheckStatus.ts
